fix(adm): fixing stats functions
